Fix misc. debug message arguments and other integer fixes for 32 bit.

+++ src/lib/cmodule/cmodule_worker.c
@@ -280,8 +280,8 @@ static int __rrr_cmodule_worker_loop_read_callback (const void *data, size_t dat
 		const struct rrr_msg_addr *msg_addr = data + MSG_TOTAL_SIZE(msg_msg);
 
 		if (MSG_TOTAL_SIZE(msg_msg) + sizeof(*msg_addr) != data_size) {
-			RRR_BUG("BUG: Size mismatch in __rrr_cmodule_worker_loop_read_callback %i+%lu != %lu\n",
-					MSG_TOTAL_SIZE(msg_msg), sizeof(*msg_addr), data_size);
+			RRR_BUG("BUG: Size mismatch in __rrr_cmodule_worker_loop_read_callback %llu+%llu != %llu\n",
+					(unsigned long long) MSG_TOTAL_SIZE(msg_msg), (unsigned long long) sizeof(*msg_addr), (unsigned long long) data_size);
 		}
 
 		callback_data->worker->total_msg_mmap_to_fork++;
